Do not install wpa_supplicant.conf
It was referenced (as a command-line option) by our downstream D-Bus
service files, but not by the upstream versions, and not by either
version of the systemd service files. It's one less thing to think
about, and fractionally less I/O during boot as well.

We don't really need the ctrl_interface Unix socket, or any other
configuration, because connman talks to wpa_supplicant via its D-Bus
interface, which is orthogonal to the ctrl_interface.