Revert "Import upstream version 2.0"
This is not how we manage source in git trees. The source from upstream has to be imported in the upstream branch using gbs or by pointing to a remote git, ONLY local changes should go to the tizen branch. 

Please do it properly and rebase tizen to remove this offending change.

Follow the instructions on

https://source.tizen.org/documentation/reference/git-build-system/upstream-package

This reverts commit 649d66a9ed4a477b136137cc77183eeaad128fcf

Change-Id: I8578be38cb135716f9ff2f18b5c3948bdd7b9857