D-Bus: Add DeviceFoundProperties signal for discovered peers
This signal sends the peer properties to applications for discovered
peers. The signature of this event is "oa{sv}". This event is needed
because the current DeviceFound signal provides only the peer object
path. If there are many peers in range there will be many DeviceFound
signals and for each DeviceFound signal, applications would need to use
GetAll to fetch peer properties. Doing this many times would create
extra load over application as well as over wpa_supplicant, so it is
better to send peer properties in the event so that applications can
extract found peer information without extra steps.

The existing DeviceFound signal is left as-is to avoid changing its
signature.

The issue is not applicable to the control interface because the
P2P-DEVICE-FOUND event includes peer info in it, but over D-Bus
interface DeviceFound provides only the peer object.

This is taken from upstream wpa_supplicant code:
http://w1.fi/cgit/hostap/commit/?id=82b9ec3125d628fdf1773b1d1184fbc1579dc3c8

Change-Id: I3329d479c6549125dfa2fdd96e9f53dd5d5888db
Signed-off-by: Nishant Chaprana <n.chaprana@samsung.com>